Migrate to decimal

+++ src/bank/calculate/balance.rs
@@ -1,5 +1,6 @@
 // See how much moneys the bank has
 
+use bigdecimal::BigDecimal;
 use diesel::{Connection, MysqlConnection, RunQueryDsl};
 
 use crate::database::tables::bank::BankInfo;
@@ -11,12 +12,12 @@ impl BankInterface {
     /// See how much money is in the bank.
     /// 
     /// Returns error if we cant get the bal
-    pub(crate) fn get_bank_balance(conn: &mut MysqlConnection) -> Result<i32, diesel::result::Error> {
+    pub(crate) fn get_bank_balance(conn: &mut MysqlConnection) -> Result<BigDecimal, diesel::result::Error> {
         go_get_bank_balance(conn)
     }
 }
 
-fn go_get_bank_balance(conn: &mut MysqlConnection) -> Result<i32, diesel::result::Error> {
+fn go_get_bank_balance(conn: &mut MysqlConnection) -> Result<BigDecimal, diesel::result::Error> {
     conn.transaction(|conn|{
         let the_bank: BankInfo = bank.first(conn)?;
         Ok(the_bank.doints_on_hand)
