+++ src/event/implementations/ubi_time.rs
@@ -1,5 +1,6 @@
 // UBI dispersal call
 
+use bigdecimal::BigDecimal;
 use diesel::MysqlConnection;
 
 use crate::{bank::bank_struct::BankInterface, event::event_struct::EventCaller};
@@ -8,7 +9,7 @@ use diesel::result::Error;
 // Collect taxes
 impl EventCaller {
     /// Collect taxes as defined in the bank.
-    pub(crate) fn ubi_time(conn: &mut MysqlConnection) -> Result<Option<u32>, Error> {
+    pub(crate) fn ubi_time(conn: &mut MysqlConnection) -> Result<Option<BigDecimal>, Error> {
         // Call it
         BankInterface::disperse_ubi(conn)
     }
